Adds lots of new features
Adds an ID and/or Class when creating an element.
Can pass objects when creating an element.
Can add multiple classes with a single `Class()` method by opassing an array.
Can pass an Object to the `Data()` method with key value pairs
Set multiple attributes at once with the `Attr()` method, which takes an Object.
Create Chains or Siblings inside an Object.